Refactor Judge utils (#18622)
