+++ tests/telemetry/test_tracked_events.py
@@ -733,16 +733,18 @@ def test_invoke_custom_judge_model(
         else:
             with (
                 mock.patch(
-                    "mlflow.genai.judges.utils._invoke_litellm_and_handle_tools",
+                    "mlflow.genai.judges.utils.invocation_utils._invoke_litellm_and_handle_tools",
                     return_value=(mock_response, 10),
                 ),
                 mock.patch(
-                    "mlflow.genai.judges.utils._invoke_databricks_serving_endpoint"
+                    "mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter._invoke_databricks_serving_endpoint"
                 ) as mock_databricks,
             ):
                 # For databricks provider, mock the databricks model invocation
                 if expected_provider in ["databricks", "endpoints"]:
-                    from mlflow.genai.judges.utils import InvokeDatabricksModelOutput
+                    from mlflow.genai.judges.adapters.databricks_serving_endpoint_adapter import (
+                        InvokeDatabricksModelOutput,
+                    )
 
                     mock_databricks.return_value = InvokeDatabricksModelOutput(
                         response=mock_response,
